requirements.txt
h5py 
pyyaml
sharedarray
tensorboard
tensorboardx
wandb
yapf
addict
einops
scipy
plyfile
termcolor
timm
torch-cluster
torch-scatter
torch-sparse
torch-geometric==2.5.3
colorhash
torchmetrics